Aggregate shadow variable setup+resolution
This PR brings together all pieces to set up a shadow variable
as we start resolving a ForallStmt. The following now happens
all in one spot, spacially (implementForallIntents.cpp)
and temporarily (during compilation), for each shadow variable
of the ForallStmt:

* Detect if an outer-variable reference should be converted
  to an "implicit" shadow variable.

  - Such references are examined while traversing all SymExprs
    in the loop body.

  - If it is, create this shadow variable.

  - Redirect the SymExprs in the loop body to this shadow var
    as they (the SymExprs) are traversed (new with this PR).

* Decide whether the shadow variable should be pruned.
  If it should, then:

  - If it is an implicit shadow variable that would have been created
    for an outer-variable reference, then it is not created,
    so there is nothing to un-do (new with this PR).

  - If it is an "explicit" shadow variable, i.e. it was listed
    in the with-clause in the Chapel code, the SymExprs in
    the loop body already reference the shadow variables.
    In this case, those SymExprs are redirected back to
    the outer variable right away (new with this PR).

* Set the shadow variable's concrete intent (when starting
  with `const` or the default intent), type, qual, flags.

* Populate the shadow variable's initBlock() and deinitBlock().

* If the shadow variable has a reduce intent:
  - Add reduce-op setup/teardown before/after the ForallStmt node.
  - Create the additional reduce-op shadow variable.

* Resolve the shadow variable's init/deinit Blocks().

WHILE THERE

I removed the follow code because the condition does not fire.
The type of the shadow variable is still adjusted using
getManagedPtrBorrowType().

```C++
  // Special case for owned -- don't want ownership transfer on
  // forall intent by default
  if (svar->getValType()->symbol->hasFlag(FLAG_MANAGED_POINTER) &&
      (svar->intent == TFI_DEFAULT || svar->intent == TFI_CONST)) {
    svar->intent = TFI_CONST_REF;
    return;
  }
```

PRUNING

* What it is:

Uses of an explicit `ref`-intent shadow variable - are reverted to
references directly to the outer variable. References to an outer
variable that would be replaced by an implicit `ref`-intent shadow
variable - are not replaced.

* Why:

Pruning has existed since shadow variables were introduced because of
the complications that we would need to handle downstream in the
compiler otherwise. This PR does not address this aspect, however
it makes pruning-related code simpler.