Set different default max registered memory regions on XE and XC.
The Gemini NIC on XE systems and the Aries on XC systems differ markedly
in how they support registered memory and in particular, in how their
respective architectures limit how much memory can be registered.
Reflecting this, set different default maximum numbers of registered
memory regions on them.  As part of this, update the error message
printed when we run out, and update the documentation where we talk
about that also.