Changes to support --library-makefile compilation with ./configure set up
Michael pointed out that the prettifying changes for --library-makefile output
(which replaced all exact paths with $(CHPL_HOME) ) wouldn't be sufficient for
when Chapel was installed using ./configure.  This change adds the additional
makefile variables which will help with that.  It also copies the script used
to make these replacements when configuring Chapel - we ran into failures
otherwise