Adjust a test to avoid delete-borrow warning
This test was reporting "'delete' can only be applied to unmanaged classes"
due to #10568. Convert the being-deleted things to 'unmanaged', as was done
in #10568.

The Chapel-erific style calls for these fields to be 'owned' instead.
However, this test was added to exercise the pattern of 'delete'-ing
in the destructor, now deinitializer. So we keep it that way.