File bug against overload checking when child class is generic
I ran into an overload bug in the DataFrames prototype in which a
method marked as override was claimed to be problematic due to not
having a matching parent method; yet removing the override caused the
complaint that the override was missing (i.e., there was no way to
please the override checker).

This is my best attempt to reduce that case to a simple standalone
case.  I believe that in order for it to work, the sublcass had to be
generic.  It also seemed to depend on the argument being a 'shared'
class type (though I didn't try very hard to explore this).