Have library mode flags trigger library mode if it is not already set
Prior to this change, passing one of ``--library-dir``, ``--library-header``,
``--library-makefile`` or ``--library-python`` without ``--library`` would not
trigger a library compilation and also would not complain that the additional
flag was not set.

Since the flags are kinda long (but need to be to distinguish what operations
will happen), it seemed unfortunate to require both when those flags imply that
``--library`` is necessary.  So, have their presence turn on library compilation
if it is not already turned on.