Cleaner compiler-added shadow variables
This PR makes it easier to change the interface calls
for reduce intents and, more generally, the code to be generated
before/after a forall loop and at task startup/teardown.
The generated C code should remain unaffected.

Background: when a ForallStmt is lowered, it is replaced with copies of
ForallStmt::loopBody() and ShadowVarSymbol::initBlock()/deinitBlock().
When copying, each shadow variable is replaced with its "live"
counterpart. For example, a TFI_TASK_PRIVATE shadow variable
is replaced with the VarSymbol* created for the task-function
FnSymbol being processed. As another example, at the top level
a TFI_REDUCE_OP shadow variable is replaced by the global reduce op.

This change extendes the set of compiler-added ShadowVarSymbols.
The role of each ShadowVarSymbol is clarified, making it easy
to create the SymbolMap to be used when copying the loopBody and
initBlocks/deinitBlocks.

This also changes how the compiler generates the pre/post-ForallStmt
code to set up the global accumulation state (AS) for a reduce intent.
Now it is independent of the initBlock/deinitBlocks. This will be
particularly helpful when we switch to the new UDRI #9879.

The outerVarSym() use is clarified as well:

* For a user-defined shadow variable - TFI_IN, TFI_CONST_IN, TFI_REF, TFI_CONST_REF, TFI_REDUCE - outerVarSym() points to the corresponding user outer Symbol.

* For a TFI_TASK_PRIVATE, outerVarSym() is NULL - this shadow variable does not have an outer variable, by definition.

* For a compiler "parent" helper - TFI_IN_PARENT, TFI_REDUCE_PARENT_AS, TFI_REDUCE_PARENT_OP - outerVarSym() points to the corresponding top-level aka "global" variable.

* For a TFI_REDUCE_OP, outerVarSym() is NULL - it is unused.

Adjust cleanupVoidVarsAndFields() to handle the case where the
accumulation state is void. This happens with the current
implementation of partial reductions, as tested by:
reductions/partial/DR-with-shaped-exprs
reductions/partial/DRtest
npb/cg/bradc/cg-sparse-partred

Move newExtraFormal() and argIntentForForallIntent() closer to their
since place of use; simplify and update them.

While there, change a few functions from extern to 'static' in
implementForallIntents.cpp.

Enacted "nicknames" for ShadowVarSymbol* variables within the compiler
For example, "SI" is a ShadowVarSymbol* for an in/const in-intent
shadow variable, "RP" is for a Reduce oP ShadowVarSymbol*, PRP is for
a Parent RP ShadowVarSymbol*, etc. These are documented in the
ForallIntentTag enum definition.

Introduce expandShadowVarForall(). It controls the new shadow variables
that are added to a ForallStmt within the parallel iterator.
While such a ForallStmt will undergo (later) its own lowering, too,
the difference is that the newly-added shadow variables reflect
forall intents from a different ForallStmt. Forall intents for the
same ForallStmt are handled in expandShadowVarTopLevel().