Replace the main module error message in library compilation
Prior to this change, a program with multiple modules and no main function would
generate the same error message when compiled normally and for use as a library,
recommending the addition of a main function or the use of the `--main-module`
flag.  However, in library compilation "main" has no special and will generate
warnings if present.

We do use the "main" module to determine the name of the library if no explicit
name is provided, so it is a problem to not know what to call the library.  This
change makes it so that when a library is compiled with multiple modules and no
-o name is specified, the user will get an error that asks for the -o flag (and
does not mention "main").