Lock rf_done pool elements, not pools.
Instead of locking entire pools of rf_done entries and then allocating
an entry out of that pool, lock individual entries to allocate them and
unlock them to free them.  Also, change how we traverse the pool, such
that a given thread attempts to use any given entry as infrequently as
possible.

This work is part of an attempt to repair the FFT performance regression
that resulted from dynamic heap registration (PR 7700).  It seems as
though it ought to have minimal overhead, but we find nevertheless that
FFT, especially at higher sizes (n=24, say) doesn't perform much better
with this code than it does on master, while it does perform much better
with the old stack-based rf_done indicators than it does one master.  So
this work, if it survives, will be an adjunct to bringing stack-based rf
done indicators back, rather than the primary solution to the FFT
regression.