Improve our task counting for remote tasks
For code like `coforall loc in Locales do on loc` we used to generate the
on-stmts wrapper functions as something like:

```c
void wrapon_fn() {
  chpl_taskRunningCntInc();     // increment local running task counter
  on_fn(...);                   // actual on function
  _downEndCount(endcount, nil); // decrement remote end count
  chpl_taskRunningCntDec();     // decrement local running task counter
}
```

Where the issue is that we decrement the remote end count before decrementing
the local running task count so the coforall sync can finish and new remote
tasks can be started before the previous ones are completely finished. In
practice I think this will only happen for local oversubscription testing. This
just adjusts our so that we insert the task counting adjustments directly
before and after the actual on fn call (previously we inserted them at the
beginning and epilogue of the wrapper function. Technically the actual task is
still running, but it's about to end so for the purposes of task counting, I
think it's the right call to decrement the local counter before we notify the
coforall sync that we're done.

We now generate something like:
```c
chpl_taskRunningCntInc();     // increment local running task counter
on_fn(...);                   // actual on function
chpl_taskRunningCntDec();     // decrement local running task counter
_downEndCount(endcount, nil); // decrement remote end count
```

Which allows the repeated coforall test to pass (I ran 20,000 trials)