Add helper function to indicate whether a FnSymbol was user-defined
This will also help with my "resolve concrete functions effort" as
it continues to get fleshed out.  My experience is that lots of
things that are compiler-defined are not labeled as such.  I think
the approach here is first to expand the list of flags that cause
this new method to return false; and then to look for where the
functions with those flags are created and add the compiler-defined
flag to them...