fixpath.py: Explicitly handle unset CHPL_HOME or env vars
Using a default value of ' ' for path when MANPATH was unset led to
changing the user's MANPATH to include a component of ' ', if
CHPL_HOME was set.  Just return the empty path in this case (the path
is already empty, and we're only supposed to be removing things from
it anyway) instead of trying to run the normal path splitting and
filtering in such a way as to end up a no-op.