Fix problem with generic class overriding concrete
If the receiver is generic, we were getting override errors no matter
what, and it tried to instantiate with the parent function details, but
that doesn't help since the receiver type can't come from the parent
function - it can only come from the call site.

Fix it by explicitly handling the case in which the receiver type is
generic.

Add tests for this pattern and un-future the test added in #10707.