Refactor checks to squash compiler-defined overrides for end-users
Move these into the checkOverrides() routine and refactor slightly
with what's there.  Ultimately, I think this routine will want more
refactoring so that we only complain about internal modules (and
maybe standard and package modules) when developer is on, but this
approach lets us keep our code up to snuff for now.