Preserve the shape of for-expressions
This is the counterpart of #9994 for (serial) for-expressions.

Next step: enable shape preservation for both for- and forall-
expressions when the shape is a range. The challenge there is
that there is no domain object that is already created. So
a new domain needs to be created from the range.

IMPLEMENTATION NOTES

In addition to the _shape_ field added in #9994, each _iteratorRecord
that represents a for-expression also gets the _fromForExpr_ method.
Which is queried in the shape-ful chpl__initCopy() to determine whether
the being-copied-from _iteratorRecord should be traversed in a for-
vs. forall-loop.

To propagate this within the compiler, PRIM_ITERATOR_RECORD_SET_SHAPE
calls now have a argument telling whether the IR is coming from
a for-expressions.

While there, removed some now-obsolete comments in ChapelArray.chpl.

TEST CHANGES

* knucleotide stores a for-expression in an array 'arr' then sorts it.
This used to work because the array was created in a "shapeless"
manner, i.e. 1-D default-rectangular. Now the array is associative
because freqs.domain - the domain the for-expression iterates over
- is associative. Associative arrays currently cannot be sorted
with quickSort(). So instead iterate over arr.sorted().
The sorted() iterator creates another temp array. This is fine
because 'arr' is very small.

* The submitted knucleotide is now .notest because of the above.

* AverageFiles.chpl started preserving the shape of a **promoted**
expression - good. I am surprised because it was is not the goal
of this PR. In the following simplification of the test, `S` used
to be shapeless, now has the domain D:
```chpl
var image: [D] real;
var S = (image / 45) -  image;
```