Handle autoDestroy of void shadow variables
The tests with user-level implementation of partial reductions
have reduce accumulation state of void type. This results
in insertion of autoDestroys with void arguments, since #10724.
cleanupVoidVarsAndFields() then remove that argument, making it
a 0-argument call. This was unexpected and caused a crash in the
parallel pass when compiling reductions/partial/DR-with-shaped-exprs
under numa.

This PR extends cleanupVoidVarsAndFields() to remove such 0-arg
autoDestroy calls.

It would be better not to add the autoDestroy call in the first place.
However, when this is done in implementForallIntents.cpp /
insertDeinitialization(), we do not know the types yet.

While there, remove commented-out code left over from #10724.
Also, during lowerForallStmtsInline(), do not insert a DefExpr
for shadow variables of void types.

This resolves the failure in reductions/partial/DR-with-shaped-exprs
in nightly numa testing.