Remove rf_done locking for most worker threads.
Re-partition the remote fork 'done' pools so that the first maxTaskPar
threads each have private partitions and don't need to lock, while all
other threads share the final partition and do have to lock.  To support
allocating 'done' indicators, to the existing 0 and 1 ('not done' and
'done') values we add -1 to mean 'free'.  Also, make the done indicators
8 bytes long instead of just 1, to reduce cache false sharing effects.
This all removes a bit more overhead from the allocate and free paths.