Simplify rfDone pools.
This is just a final change in which we create a single pool of rfDone
indicators and partition that into thread-private chunks for the worker
threads and a final chunk shared by everyone else.  The only real effect
is to simplify the array indexing for the pool.  It doesn't seem to make
the performance any better.

At this point, based on 10 XC-16 runs with performance settings, it
appears FFT performance is a wash between stack-based rfDone indicators
and this improved pooling code.