Initializes sizesPerDim only if usePollyArrayIndex
PR #10727 added sizesPerDim to DefaultRectangular.  This field is only
used in an experimental Polly integration, which is enabled (for now)
with the config param usePollyArrayIndex.

After PR #10727, the test
`optimizations/remoteValueForwarding/serialization/domains`
failed due to an increase in communication when creating a
DefaultRectangular array. The new communication comes from
initializing sizesPerDim.

If performance testing indicates it to be important, we will additionally
make sizesPerDim a void field in another PR.