Add checks at resolution for exporting functions that take or return strings
The C type definition for our notion of strings is not visible at the moment,
due to not being able to export type definitions in Chapel yet.  But prior to
this change, compiling a library with exported functions involving strings did
not cause an error - the user would discover they had no definition of the type
when they tried to link to the compiled library.  This change adds a compilation
error specifically for strings, so that users are not surprised at library link
time and instead know right away that they should use c_strings for the time
being.