Add require output for libraries to the generated makefile
Seems to work fine, but probably doesn't cover .h files when a type specified
there is used as an argument or return type for an exported function.  Not sure
if something more needs to be done for .c and .o files specified in this way,
probably a question for Brad (my initial guess is no, that they are just
packaged into the library we created and so it is fine).