Add a new dsiReallocate that is aware of allocated size and actual array size.
When array.pop_front and friends caused a reallocation, they were only paying
attention to the allocated size, which resulted in copies of elements that
were not going to be in the newly resized array.  Make dsiReallocate use both
the allocated bounds and the logical array bounds.  Allocate and free based
on the allocated bounds, but copy elements based on the logical array bounds.

This fixes a leak in the newly added test from issue #10520.