Try a different hybrid waitFor approach
Previous version just did 1,000 remote reads before switching to an AM. That
suffered at scale (tested up to 256 nodes) because we overwhelmed as single
node with a lot of network traffic. This approach always does remote reads, but
it sleeps in between trials. This may help with, but we'll have to do some
tuning on the sleep amount and other parameters.