Cleanup NIC AMO spacing/formatting
Move amo perfstats_inc to a common location, and unify alignment of some
assignment statements for setting up the post AMO post descriptors.

This switches the NF operand casting to use a ternary operator and more
importantly corrects the operand casting for the buffered NF case. Previously,
we always casted to a 64 bit type, which was wrong for 32 bit operations.