Futurize the test of taking/returning a chpl_external_array explicitly
The behavior I needed is present (generated chpl_external_array instances get
converted appropriately while explicit instances are left alone), but you cannot
build a Python module with what is currently generated for this case.  We need
to generate more in the .pyx file, but while it would not take much time to do
so, I think I need to move on to other things.  So, save what I expect is needed
into the .good file and futurize the test