Partial update of tests
Add a prediff which runs a python file that used the built module for each test,
and a COMPENV file which extends the PYTHONPATH environment variable to include
the generated lib/ directory where all the necessary files live.  uintFunctions
actually has been given the python file in question, the other tests require
theirs to be written.  uintFunctions also has had its catfiles removed and its
.good file updated.  Its prediff is in an interim state - it currently builds
the module using the generated files, instead of relying on the compiler to
perform that task.  I hope to have the compiler perform that task in the next
step (which is why the other prediffs don't include that)