Remove iteratorLeaderMap, iteratorFollowerMap
These two maps were used to find the leader/follower iterators for
promoted expressions. To reduce the number of code paths in the compiler,
we switch to finding the leader/follower iterators using regular
call resolution. For that: (a) make those iterators' names unambiguous
and (b) do not flag them with FLAG_INVISIBLE_FN.

While there, reformat a function for improved readability.

When searching the initializer call within a promotion wrapper for
an initializer, expect that call to be named "_new". This will not
work when promoting an (old-style) constructor call, because such
occasions will be disallowed soon.

Update .good/.bad files for the new names of the internal
iterator-record types.