build_configs/cray: Add setenv-xc-x86_64.bash for Chapel Cray XC modules
New script file setenv-xc-x86_64.bash builds the Chapel Cray XC module
for x86_64.

It is based on setenv-example-3.bash from the parent directory, and the
basic structure is the same. However, many details were added to successfully
build a complete Cray module with Python-venv, chpldoc, mason, 192 runtime
configs, etc.

It is relocated to this build_configs/cray-internal subdir to indicate
this is a Cray-specific application of the build_configs/setenv pattern.