Make LLVM work with dynamic linking on a Cray XC system (#10902)
Make LLVM work with dynamic linking on a Cray XC system (#10902)

[reviewed by @ronawho ]

Certain libraries must be explicitly specified for dynamic linking to work when using the PrgEnv compiler/linker wrapper. This was specified in Makefile.cray-prgenv which was not accessible to the LLVM back end. This change moves the required logic out to a shell script so it can be accessed both directly by a PrgEnv compilation, and indirectly for the benefit of an LLVM compilation.