Sanity-check hugepage-related env vars when CHPL_COMM=ugni.
Print a warning if we don't seem to be using hugepages, because of the
potential performance loss.  Also print warnings if we do seem to be
using hugepages but HUGETLB_NO_RESERVE or CHPL_JEMALLOC_MALLOC_CONF are
not set, since those are also important for performance and their
absence implies that the Chapel launcher (which always sets them) was
not used.

Note that we should really print an error instead of a warning if
CHPL_JEMALLOC_MALLOC_CONF is set and specifies a chunk size larger than
the hugepage size, or is not set and jemalloc's default 2 MiB chunks are
larger than the hugepage size.  In those cases we can get misaligned
chunks, which will lead to segfaults or internal errors, so we're better
off just erroring out.  This is work for the future, though, because
getting all the needed information together early enough is going to
require some reorganization which I don't have time for right now.

As a side effect of the current effort I moved jemalloc_conf_ev_name()
from comm-ugni-launch.c to comm-ugni-heap-pages.h and made it an inline
static called chpl_comm_ugni_jemalloc_conf_ev_name(), because now I
needed the name of the jemalloc configuration env var in both the
launcher and the runtime proper.  Really it probably ought to be in a
file related to the jemalloc memory layer, but getting it there for the
launcher ended up looking like more work than I wanted to do just a
couple of days before feature freeze, so that will have to wait.