Stop make(1) explicitly when the compileline utility fails.
The Qthreads build needs some of the runtime build -I and -D options,
which we get with the compileline utility.  But that can fail to run
correctly; the most common reason has been that one of the things it
needs to reference hasn't been built yet.  Such a failure doesn't halt
the make since compileline is run in a make 'shell' function invocation
to set a make variable, not in a target build command.  The make will
eventually fail, due to an error compiling some Qthreads source file
because the -I or -D options it needs aren't there.  But the root cause
is buried much earlier in the log, where the compileline invocation
failed, and the connection between the two isn't at all obvious.

We've solved the instances of this we've seen, by adding prerequisites
reflecting compileline's dependencies, but we may not have seen all of
them and certainly new dependencies lacking corresponding prerequisites
could be created in the future.  So here, force the make to halt right
away when the compileline invocation fails.