Add a compiler error for using '.type' on something that is already a type
Applying the .type operation to something that is already a type now results
in a compiler error instead of getting ignored in most cases.  This is handled
during preFold for PRIM_TYPEOF.  It needs to be handled during resolution so
that it can apply to uses of '.type' on function calls that return types.

Updated test/arrays/ferguson/runtime-type-type.good to expect the error
message instead of silently ignoring the .type operation. Removed the .future
because it now generates a nice error instead of an unresolved call error
about an internal function.

Added a test that tries to use '.type' on some simple types, and another that
uses '.type' on a function call that returns a type.