Fix exists and isLink to not error when the empty string is passed
Prior to this change, exists and the functions that depend on it would throw a
bad access system error when passed "" as their argument.  isLink would do the
same.  However, Python does not consider this cause for an error and instead
just returns false - we've tried to base our behavior off of Python as much as
possible for these functions.  This change unifies these functions in line with
Python.