Ensure _RuntimeTypeInfo is initialized
The array created out a for/forall/promoted expression used to
be created out of uninitialized _RuntimeTypeInfo in some cases
when the array's element type is a runtime type.

That is because the _RuntimeTypeInfo requires having at least
one element already available. Whereas the array is created
**before** the expression is executed, so there are no elements yet.

This change create stand-in _RuntimeTypeInfo to use when creating
the array. For example, if elements are domains, the _RuntimeTypeInfo
results in an empty domain. If elements are arrays, the _RuntimeTypeInfo
is for an empty array. And so on, recursively.

When the elements are/contain non-DefaultRectangular arrys/domains,
the compiler issues an error.

This is a workaround for invalid memory accesses,
while waiting for a worked-out design solution.