Shape preservation with ranges
This extends the approach of #9994 and #10758
for shapes that are ranges.

The main challenge was to ensure no memory leaks **and**
no double-deletions. For that, the cases where the shape
domain already exists and where the shape domain needs
to be created out of a range are handled differently.
The remainder of chpl__initCopy() for these two cases
is factored into chpl__initCopy_shapeHelp().

Update a test that depended on the old behavior.
Added more checking to the prototype partial-reduction implementation.
While there, updated a comment.

Requires #10941 to ensure valid memory accesses.