ensure _RuntimeTypeInfo is initialized (#10941)
The array created out a for/forall/promoted expression used to
be created out of uninitialized _RuntimeTypeInfo in some cases
when the array's element type is a runtime type.

That is because the _RuntimeTypeInfo requires having at least
one element already available. Whereas the array is created
**before** the expression is executed, so there are no elements yet.

This change creates stand-in _RuntimeTypeInfo to use when creating
the array. For example, if elements are domains, the _RuntimeTypeInfo
results in an empty domain. If elements are arrays, the _RuntimeTypeInfo
is for an empty array. And so on, recursively.

The stand-in mirrors the original without accessing any of the original's
run-time contents. Given that the latter are uninitialized.

Background for the above: an array or domain type in Chapel code
is represented at run time by a _RuntimeTypeInfo record instance.
Declaring a variable or such a type involves a run-time conversion
from _RuntimeTypeInfo to an instance of _array or _domain. (And
conversely, taking .type of such a variable creates a _RuntimeTypeInfo
instance.) Therefore when _RuntimeTypeInfo is uninitialized,
creating an array/domain out of it results in invalid memory accesses.

When the elements are/contain non-DefaultRectangular arrys/domains,
the compiler issues an error. This is because creating a stand-in takes
more effort. While the other DefaultDist cases can be added easily,
such as enum, associative, etc., a stand-in for a Block domain or array
is more involved - both in terms of implementation effort and execution-time
cost.

This is a workaround for invalid memory accesses.
Proper design and solution may replace this workaround.

r: @benharsh