Make performance tests make better use of managed classes
While running release-over-release testing, a number of performance
tests fail on old versions of the compiler due to using 'owned',
'unmanaged', etc.  This PR represents an attempt to (a) use
undecorated 'new's on performance tests in hopes of making them more
backwards compatible and (b) using more principled ownership types to
clean up the implementations and make them more representative of the
type of Chapel we'd like to be writing going forward.

This does not hit all such performance tests.  Some key missing ones
are: LCALS, CoMD which were bigger than I had time to take on in this
pass.