Don't allow isSubtype/<= between type and value
Because normalization of type query arguments relies on
a PRIM_IS_SUBTYPE with a mix of type and values, this PR adds
a new primitive, PRIM_IS_SUBTYPE_ALLOW_VALUES, which has the
old behavior, and uses it for type query arguments and legacy :
in where clauses.

Some module code updated to use .type to get the appropriate type.

exprAnalysis doesn't handle PRIM_IS_SUBTYPE anymore, since it
should be removed by the time it runs.