Reduce the amount of memory used by oversubscribedArrayAlloc
This test allocates large arrays in parallel to stress ugni's dynamic array
allocation/registration code. Previously, it was using 1/3 of physical memory,
but this drops it to use 1/10.

We've been seeing sporadic OOMs on ARM nodes that have severe fragmentation and
dropping to 1/10 mem helps quiet testing. All we really care about is that each
array is large enough to trigger the ugni code path, and we're still well above
that on any node type that I know of.