Fix leak for default new with --no-warnings
PR #10957 disabled the warning for `new SomeClass` if --no-warnings
is provided. But, it had an error that caused the allocation to leak
in that situation, even though it should be equivalent to
`new borrowed SomeClass`.

This PR fixes that problem and adds a test to catch it if it ever
happens again.