Always prepend "lib" to libraries created using --library
Prior to this change, we would always prepend "lib" to libraries created without
the `-o` flag.  Libraries created with the `-o` flag would use exactly the name
given (plus the relevant extension for static or dynamic code).  This was fine
for using these libraries with C, but Cython expected every library to start
with "lib", especially when compiling statically.

To avoid confusion in --library-python compilation (and follow the unwritten
rules about libraries), we decided to update compilation with `-o` to also
prepend "lib" to the name.  This does mean that the chapel compiler will never
create a foo.a/foo.so file but the user can always rename the generated library
after the fact if they really care that much.

Note that this means that `-o libfoo` will generate liblibfoo.a/liblibfoo.so
for now.  Will fix in next commit