Fix performance regression in pop_front/pop_back from PR #10793
PR #10793 fixed a memory leak by tracking both the allocated memory and
the logical array size within dsiReallocate.  It did this by creating an
extra domain from an extra tuple of ranges for the logical size.

This extra domain added some performance overhead when reallocating arrays
for pop_front and pop_back.  Remove the overload of dsiReallocate that was
added in #10793.  Add a new overload (for rank==1 only) that has range
arguments instead of tuples of ranges, and avoids creating the extra domain.

This should fix the performance regression while continuing to avoid the
memory leak.