Handle no-hugepage-module better in hugepage-sanity test.
The hugepage-sanity test was failing in Cray XE testing with an internal
comm=ugni error due to a memory registration error.  It would behave as
expected if we had the craype-hugepages2M module loaded but if we had
the craype-hugepages16M module loaded, the guaranteed-registered memory
area could not be registered.  In that case the region was at a 2M
boundary because that is the default kernel hugepage size.  Our guess is
that some job-level setup done by the system launcher (on the login
node, before the _real wrapper effectively unloaded the module) caused
the NIC to expect 16M pages, so a 2M-aligned memory region could not be
registered.  (Memory management support is quite different in the Gemini
and Aries NICs, so seeing different behavior on XE and XC systems for
this admittedly oddball test is not particularly a surprise.)

The solution, implemented here, is to double-wrap the test, that is, to
wrap both the launcher and the _real.  The 1st execopts case, which
tests for the module not being loaded, is now set up in the launcher
wrapper.  Setting up for the 2nd and 3rd execopts cases stays in the
_real wrapper.