Task-private variables in the spec
Add the specification of task-private variables.

Given that the with-clause and its syntax \sntx{task-intent-clause}
are defined for task intents i.e. in the Task Parallelism chapter,
update the corresponding pieces there as well.

Introduce the term "shadow variable" as the thing that is referenced
implicitly in the forall loop body.

I put TPVs in their own subsection of the spec because I felt
that the reader will find them more easily that way, than if
if they were defined inside the "Forall Intents" subsection.