Revert changes to PTRANS
Elliot pointed out that it was weird that previously this had used
'unmanaged' yet had no 'delete's and didn't leak.  While this passed
testing on my Mac, it caused crashes and valgrind failures on linux64
due to (essentially) a double-free.  This suggests to me that we
don't have memory management correct for distributions yet.